feat: enable launching browser with ws ports and attaching
This ports the necessary bits of discovery and connection logic from
js-debug to allow attaching to a websocket inspector endpoint. Then,
it uses that to allow launching browsers with a debug port (rather
than simply a pipe) as well as attaching to already-running browsers.

It also fixes an issue where a timeout happens when trying to close
a remote browser.

Refs: https://github.com/microsoft/vscode-js-debug/issues/640
Refs: https://github.com/microsoft/vscode/issues/104464