fix: adopt improved connection process for WSL
@aeschli mentioned that he saw occasional port issues where WSL ports
don't get forwarded to Windows in Remote WSL. I realized that's the
same bug here (I initially though it was my problem, but after
investigation I realized differently.)

So I copy his approach of using a subprocess to communicate with the
port on the WSL side. If this fails we fall back to the usual port
communication, but ti seems to work pretty nicely.

Also, swap over to esbuild while I'm at it, since duplexer3 is an ESM module.